backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
faster-whisper>=1.0.0
kokoro>=0.9.2
soundfile>=0.12.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
pymupdf>=1.23.0
httpx>=0.25.0
webrtcvad>=2.0.10
numpy>=1.24.0
python-dotenv>=1.0.0
psutil>=5.9.0
